LayoutTests/http/tests/appcache/remove-cache-expected.txt

Test that a 404 response for manifest results in cache removal.

Frame 1: Manifest is still available, so a new master resource is added to the cache.
Frame 2: Manifest loading results in 404 response, so the cache group becomes obsolete, and an obsolete event is dispatched (because the document in frame was associated with a cache in the group).
Frame 3: Manifest is still 404 - the document is never associated with a cache.
Frame 4: Manifest is now available, so the document gets associated with a cache in a newly created group; the obsolete cache group is not affected.
Should say SUCCESS:

SUCCESS
